Dao implemented + logging added

+++ cafe/src/main/java/com/eugene/cafe/entity/Entity.java
@@ -2,16 +2,16 @@ package com.eugene.cafe.entity;
 
 public abstract class Entity {
 
-    protected long id;
+    protected int id;
 
     public Entity() {
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 }
